App
Two classes, Cannon & Musket, implement the Ballistics interface. They both can call methods defined in the Ballistics interface. Interfaces can define methods, but cannot contain any code. The code the method executes will be defined in the class(es) in which it is implemented. 

Here an instanced object of both Cannon & Musket call methods implemented by the interface. You can also use an interface to create a new instance of a class, such as Cannon. You can then use that variable to call methods of the object type, in this case Cannon. You can also create a variable (ballistics 2 in this case) using an interface that points to a different variable, in this case the variable of an object of class Cannon (can1). As with the previous example, you can now use the variable ballistics2 to call methods defined only in the Ballistics interface, and implemented in class Cannon.


Further down a static method is built that draws upon the Ballistics interface. The method can only call methods defined in the Ballistics interface. In the example only showInfo is used.

Further above, the ouputInfo method previously mentioned is used to call the showInfo method of interface Ballistics using the variables of previously created objects (whos classes implement the Ballistics interface) to call upon the showInfo method.